fix: Return a 401 instead of a 302 when not authenticated.
Some clients will run into problems if we return a 302 here,
if they automatically follow redirects.  Consider being
logged in to some service, leave the window open for a bit,
then kube-auth-proxy automatically logs you out.  You come
back to this existing window, and click on a button that
results in an AJAX request.  If the client code is using
the WHAT-WG fetch API, they'll probably automatically
follow the 302 to the login page (which will be a 200),
and then try to pass the login page through `JSON.parse()`,
which ends in sad faces all around.  By returning a 401
error here, we make it more likely the client will do
something sensible in this situation.